add picture folder, edit header.html add admin functionality - edit and delete article

+++ TravelGuide/src/main/java/softuniBlog/controller/ArticleController.java
@@ -2,6 +2,8 @@ package softuniBlog.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
@@ -18,10 +20,16 @@ import softuniBlog.repository.UserRepository;
 @Controller
 public class ArticleController {
 
-    @Autowired
+    private final
     ArticleRepository articleRepository;
-    @Autowired
+    private final
     UserRepository userRepository;
+
+    @Autowired
+    public ArticleController(ArticleRepository articleRepository, UserRepository userRepository) {
+        this.articleRepository = articleRepository;
+        this.userRepository = userRepository;
+    }
 
     @GetMapping("/article/create")
     @PreAuthorize("isAuthenticated()")
@@ -54,6 +62,14 @@ public class ArticleController {
         if(!this.articleRepository.exists(id)){
             return "redirect:/";
         }
+        if(!(SecurityContextHolder.getContext().getAuthentication() instanceof AnonymousAuthenticationToken)){
+            UserDetails principal = (UserDetails) SecurityContextHolder.getContext()
+                    .getAuthentication()
+                    .getPrincipal();
+
+            User user = this.userRepository.findByEmail(principal.getUsername());
+            model.addAttribute("user", user);
+        }
         Article article = this.articleRepository.findOne(id);
         model.addAttribute("view", "article/details")
                 .addAttribute("article", article);
@@ -68,6 +84,11 @@ public class ArticleController {
             return "redirect:/";
         }
         Article article = this.articleRepository.findOne(id);
+
+        if(this.isUserAuthorOrAdmin(article)){
+            return "redirect:/article/" + id;
+        }
+
         model.addAttribute("view", "article/edit")
                 .addAttribute("article", article);
         return "base-layout";
@@ -82,6 +103,11 @@ public class ArticleController {
         }
 
         Article article = this.articleRepository.findOne(id);
+
+        if(this.isUserAuthorOrAdmin(article)){
+            return "redirect:/article/" + id;
+        }
+
         article.setTitle(articleBindingModel.getTitle());
         article.setContent(articleBindingModel.getContent());
 
@@ -99,6 +125,11 @@ public class ArticleController {
             return "redirect:/";
         }
         Article article = this.articleRepository.findOne(id);
+
+        if(this.isUserAuthorOrAdmin(article)){
+            return "redirect:/article/" + id;
+        }
+
         model.addAttribute("view", "article/delete")
                 .addAttribute("article", article);
         return "base-layout";
@@ -112,9 +143,23 @@ public class ArticleController {
             return "redirect:/";
         }
 
+        Article article = this.articleRepository.findOne(id);
+
+        if(this.isUserAuthorOrAdmin(article)){
+            return "redirect:/article/" + id;
+        }
+
+
         this.articleRepository.delete(id);
         return "redirect:/";
 
+    }
+
+    private boolean isUserAuthorOrAdmin(Article article){
+        UserDetails user = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        User userEntity = this.userRepository.findByEmail(user.getUsername());
+
+        return userEntity.isAuthor(article) || userEntity.isAdmin();
     }
